+++ CDJSVis/visclibs/bond_order.h
@@ -1,0 +1,12 @@
+
+int bondOrderFilter(int, int*, int, double *, double, double, double, int, double *, double *, double *, double *, double *, int *, int, double *, int);
+
+struct AtomStructureResults
+{
+    double Q6;
+    double Q4;
+    double realQ4[9];
+    double imgQ4[9];
+    double realQ6[13];
+    double imgQ6[13];
+};